LOGIN AND REGISTER PROCESS WORK WITH NO ISSUES (at least none that I could find)

+++ app/src/main/java/com/example/republicsavingsapp/CategoryDAO.kt
@@ -10,5 +10,5 @@ interface CategoryDAO {
     suspend fun addCategory(category: Category): Long
 
     @Query("SELECT * FROM categories WHERE userId = :activeUserID ORDER BY categoryName ASC")
-    suspend fun getAllForUser(activeUserID: Int): List<Category>
+    suspend fun getAllForUser(activeUserID: Long): List<Category>
 }
