+++ app/src/main/java/com/example/republicsavingsapp/Expenses.kt
@@ -7,7 +7,7 @@ data class Expenses(
     @PrimaryKey(autoGenerate = true)
     val expenseID: Int = 0,     // not needing to show the user
 
-    val userId: Int,       // the row number of the user in the Users Table (1 user:Many expenses relationship)
+    val userId: Long,       // the row number of the user in the Users Table (1 user:Many expenses relationship)
     val expenseName: String,
     val expenseDescription: String? = null,         // descriptions are allowed to be null
     val expenseAmount: String,
